# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AprovaFácil API",
    "description": "API para sistema de estudos e preparação para concursos",
    "version": "1.1.0",
    "contact": {
      "name": "AprovaFacil Team",
      "email": "contato@aprovafacil.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desenvolvimento"
    },
    {
      "url": "https://api.aprovafacil.com",
      "description": "Servidor de produção"
    }
  ],
  "paths": {
...
    "/conteudo/filtrado": {
      "get": {
        "operationId": "getConteudoFiltrado",
        "summary": "Obter conteúdo filtrado",
        "description": "Retorna conteúdo filtrado por parâmetros",
        "tags": [
          "Conteúdo"
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "description": "Tipo de conteúdo",
            "schema": {
              "type": "string",
              "enum": [
                "apostila",
                "flashcard",
                "questao"
              ]
            }
          },
          {
            "name": "concursoId",
            "in": "query",
            "description": "ID do concurso",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conteúdo filtrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}